Modified the code to make dynamic_eg.c more kernel style
Add Code:
(1) dynamic_eg.c
	(a)FUnction IOCTL: copy_from_user &  copy_to_user

Note:

Using structure as the structure to exchange data is a bad idea
This way is not good because the buffer should be continuous memory
block,hence if the buufer is overflow the data what we stored could
not know.

e.g:
struct buffer {
	char kernel_buffer[16];
	char usr_buffer[16];
};

What if the kernel buffer stored 18 byte string that means
when we get data would get part of it, the data would
be wrong.